pyproject.toml
[project]
name = "shallow-dive-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.2.0",
    "tavily-python>=0.3.3",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
